requirements.txt
streamlit==1.42.0
PyMuPDF==1.25.3
litellm==1.61.1
python-dotenv==1.0.1